VST-02 | Possible integer overflow and dead loop

+++ contracts/Vaults/VaultStakedToGlobal.sol
@@ -166,7 +166,7 @@ contract VaultStakedToGlobal is IDistributable, ReentrancyGuard, RewarderRestric
     }
 
     function _deleteUser(address _account) private {
-        for (uint8 i = 0; i < users.length; i++) {
+        for (uint i = 0; i < users.length; i++) {
             if (users[i] == _account) {
                 delete users[i];
             }
